requirements.txt
# Base dependencies
torch>=2.0.0
numpy>=1.21.0
tqdm>=4.65.0
PyYAML>=6.0
transformers>=4.30.0
datasets>=2.12.0
tokenizers>=0.13.3

# Optimizations
optuna>=3.2.0
scikit-learn>=1.2.2
scipy>=1.10.1

# Visualization and logging
tensorboard>=2.12.0
matplotlib>=3.7.1
seaborn>=0.12.2
wandb>=0.15.4

# Testing
pytest>=7.3.1
pytest-cov>=4.1.0
pytest-benchmark>=4.0.0

# Utilities
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.3.0
pre-commit>=3.3.2

# Optional dependencies
bitsandbytes>=0.39.0  # for quantization
accelerate>=0.20.0    # for distributed training
deepspeed>=0.9.0      # for deep learning
sentencepiece>=0.1.99 # for tokenization
protobuf>=4.23.0      # for serialization
googletrans==3.1.0a0  # for translation 